env folder deleted.

+++ earnings/models.py
@@ -2,6 +2,7 @@ from django.db import models
 from django.conf import settings
 from django.core.validators import MinValueValidator
 
+
 class Earnings(models.Model):
     ROI = 'ROI'
     REFERRAL_BONUS = 'REFERRAL_BONUS'
@@ -14,6 +15,7 @@ class Earnings(models.Model):
     earnings_id = models.AutoField(primary_key=True, unique=True)
     user = models.ForeignKey(settings.AUTH_USER_MODEL, models.PROTECT)
     earning_type = models.CharField(max_length=255, choices=EARNING_TYPE)
-    amount = models.FloatField()
+    amount = models.DecimalField(max_digits=12, decimal_places=3)
+
     def __str__(self):
         return self.earning_type
